# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: AllCallAll API
  version: 1.0.0
  description: Authoritative contract consumed by the primary Web client.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /realtime/tickets:
    post:
      operationId: issueRealtimeTicket
      summary: Issue a 60-second, channel-bound, single-use WebSocket ticket.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [channel]
              properties: { channel: { type: string, enum: [chat, signaling, room] } }
      responses:
        '201':
          description: Realtime ticket issued.
          content: { application/json: { schema: { $ref: '#/components/schemas/RealtimeTicket' } } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
...
  schemas:
...
    RealtimeTicket:
      type: object
      required: [ticket, channel, expires_at, websocket_path]
      properties:
        ticket: { type: string }
        channel: { type: string, enum: [chat, signaling, room] }
        expires_at: { type: string, format: date-time }
        websocket_path: { type: string }
...